+++ .clang-tidy
@@ -14,7 +14,6 @@ Checks: >
   performance-*,
   -performance-avoid-endl,
   readability-*,
-  -readability-magic-numbers,
   -readability-function-cognitive-complexity,
   -readability-identifier-length,
   -readability-named-parameter,
